♻️ ゲーム処理を構造体に分離した
