+++ src-tauri/Cargo.toml
@@ -24,3 +24,6 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 rand = "0.8"
 lazy_static = "1.4"
+
+[dev-dependencies]
+serial_test = "1.0"
